+++ config/postcss.config.js
@@ -3,10 +3,7 @@
 module.exports = {
 	map: false,
 	plugins: [
-		require(`postcss-import`),
-		require(`postcss-assets`),
 		require(`postcss-focus`),
-		require(`postcss-will-change`),
 		require(`rfs`),
 		require(`autoprefixer`)({
 			cascade: true,